+++ Routory/Routory/Presentation/Registration/ShiftRegistration/Handler/ShiftRegistrationDelegateHandler.swift
@@ -24,11 +24,6 @@ final class ShiftRegistrationDelegateHandler: NSObject {
 
 extension ShiftRegistrationDelegateHandler: WorkPlaceSelectionViewDelegate {
     func workPlaceSelectionViewDidTapChevron(_ view: WorkPlaceSelectionView) {
-//        let vc = WorkplaceSelectionViewController()
-//        vc.onSelect = { [weak self] workplace in
-//            self?.contentView?.simpleRowView.updateTitle(workplace.workplacesName)
-//        }
-//        navigationController?.pushViewController(vc, animated: true)
         let workplaces: [Workplace] = [
             Workplace(id: "1", workplacesName: "맥도날드", category: "패스트푸드", ownerId: "owner1", inviteCode: "...", inviteCodeExpiresAt: "...", isOfficial: true),
             Workplace(id: "2", workplacesName: "쿠팡 야간", category: "물류", ownerId: "owner2", inviteCode: "...", inviteCodeExpiresAt: "...", isOfficial: false),
