+++ Routory/Routory/Presentation/Registration/WorkplaceRegistration/Handler/WorkplaceRegistrationDelegateHandler.swift
@@ -42,7 +42,25 @@ extension WorkplaceRegistrationDelegateHandler: WorkplaceInfoViewDelegate {
     }
 
     func didTapCategoryRow() {
-        print("didTapCategoryRow")
+        let categoryItems: [SelectionViewController<String>.Item] = [
+            .init(title: "음식점", icon: UIImage(named: "restaurant"), value: "음식점"),
+            .init(title: "카페", icon: UIImage(named: "cafe"), value: "카페"),
+            .init(title: "편의점", icon: UIImage(named: "convenience"), value: "편의점"),
+            .init(title: "영화관", icon: UIImage(named: "cinema"), value: "영화관"),
+            .init(title: "기타", icon: UIImage(named: "box"), value: "기타")
+        ]
+
+        let vc = SelectionViewController<String>(
+            title: "카테고리",
+            description: "근무지 카테고리를 선택해주세요",
+            items: categoryItems,
+            selected: "카페"
+        )
+
+        vc.onSelect = { selected in
+            print("선택된 카테고리:", selected)
+        }
+        navigationController?.pushViewController(vc, animated: true)
     }
 }
